# Optimizations, performance, and benchmarking