--- a/jvm/src/main/coffee/engine/core/topology/torus.coffee
+++ b/jvm/src/main/coffee/engine/core/topology/torus.coffee
@@ -6,6 +6,12 @@ Topology = require('./topology')
 { pipeline }   = require('brazierjs/function')
 { rangeUntil } = require('brazierjs/number')
 
+# Why our own custom add function?  To avoid anonymous functions getting created as part
+# of the fold below.  This bad boy will add any two numbers together, no problem.
+# -JMB 07/2017
+add = (a, b) ->
+  a + b
+
 module.exports =
   class Torus extends Topology
 
@@ -42,7 +48,7 @@ module.exports =
            @_getPatchNorthWest(pxcor, pycor), @_getPatchSouth(pxcor, pycor),
            @_getPatchNorth(pxcor, pycor), @_getPatchSouthEast(pxcor, pycor),
            @_getPatchEast(pxcor, pycor), @_getPatchNorthEast(pxcor, pycor)]
-        diffusalSum = pipeline(map(getScratch), foldl((acc, x) -> acc + x)(0))(orderedNeighbors)
+        diffusalSum = pipeline(map(getScratch), foldl(add)(0))(orderedNeighbors)
         patch.setVariable(varName, patch.getVariable(varName) * (1.0 - coefficient) + (diffusalSum / 8) * coefficient)
         return
       )