--- a/jvm/src/main/coffee/engine/core/turtle.coffee
+++ b/jvm/src/main/coffee/engine/core/turtle.coffee
@@ -201,6 +201,10 @@ module.exports =
     _optimalNSum: (varName) ->
       @getPatchHere()._optimalNSum(varName)
 
+    # (String) => Number
+    _optimalNSum4: (varName) ->
+      @getPatchHere()._optimalNSum4(varName)
+
     # (Number) => Boolean
     jumpIfAble: (distance) ->
       canMove = @canMove(distance)