--- a/jvm/src/main/coffee/engine/core/turtlelinkmanager.coffee
+++ b/jvm/src/main/coffee/engine/core/turtlelinkmanager.coffee
@@ -9,6 +9,8 @@ TurtleSet = require('./turtleset')
 
 { DeathInterrupt, ignoring } = require('util/exception')
 
+ignorantly = ignoring(DeathInterrupt)
+
 # data Directedness
 All = {}
 In  = {}
@@ -47,7 +49,7 @@ module.exports =
 
       # Purposely done after resetting the array so that calls to `TurtleLinkManager.remove` in `Link.die` don't spend
       # a ton of time iterating through long arrays that are in the process of being wiped out. --JAB (11/24/14)
-      oldLinks.forEach((link) -> ignoring(DeathInterrupt)(() => link.die()))
+      oldLinks.forEach((link) -> ignorantly(() => link.die()))
 
       return