# Benchmark - No Optimizations

BZ Benchmark (V8 Node.js v6.10.3):

- Average: 54.561 seconds
- Min:     45.171 seconds
- Max:     84.966 seconds

FireBig Benchmark (V8 Node.js v6.10.3):

- Average: 96.768 seconds
- Min:     76.092 seconds
- Max:     128.195 seconds

GridWalk Benchmark (V8 Node.js v6.10.3):

- Average: 446.069 seconds
- Min:     340.928 seconds
- Max:     581.184 seconds

Wolf Benchmark (V8 Node.js v6.10.3):

- Average: 336.694 seconds
- Min:     250.894 seconds
- Max:     455.119 seconds

# Including All Optimizations

BZ Benchmark (V8 Node.js v6.10.3):

- Average: 38.339 seconds
- Min:     35.332 seconds
- Max:     41.527 seconds

FireBig Benchmark (V8 Node.js v6.10.3):

- Average: 48.298 seconds
- Min:     47.521 seconds
- Max:     49.094 seconds

GridWalk Benchmark (V8 Node.js v6.10.3):

- Average: 276.59 seconds
- Min:     274.283 seconds
- Max:     280.762 seconds

Wolf Benchmark (V8 Node.js v6.10.3):

- Average: 282.732 seconds
- Min:     185.238 seconds
- Max:     439.836 seconds