@SethTisue 2017-07-04
the spread between the min and max numbers seems alarmingly high for some of these? curious about that, wondering if those numbers are meaningful