@LaCuneta 2017-07-05
@SethTisue That's a good question, I'll have to look into that!  It could be something as simple as processes on my workstation interfering with the results as they ran.  If the random seed isn't set the same, that'd also cause big differences, but I'm pretty sure it is.  I can also get SpiderMonkey setup and check if there are similarly wide gaps between min/max using it.  

I did run manual stopwatch tests as well in the browser to confirm the results, and did see improvements that came close to the averages here.  I did not see the big differences in min/max, though, when doing it that way.